Rename `structure` into `meta`
`structure` could be misleading and conflicting
