+++ votepy/__init__.py
@@ -2,4 +2,4 @@ import votepy.rules.chamberlin_courant
 import votepy.rules.greedy_monroe
 import votepy.rules.k_borda
 
-from votepy.structure.listing import algorithms, rules
+from votepy.meta.listing import algorithms, rules
