+++ votepy/meta/structure.py
