+++ votepy/solve.py
@@ -1,7 +1,7 @@
 from typing import Callable, Iterable, Union
 from votepy.ordinal_election import OrdinalElection
 from votepy.algorithms.base_algorithm import BaseAlgorithm
-from votepy.structure.structure import get_algorithm, get_default_algorithm, get_implementation
+from votepy.meta.structure import get_algorithm, get_default_algorithm, get_implementation
 
 
 def solve(rule: Union[Callable, str], voting: Union[OrdinalElection, list[int]], size_of_committee: int, *rule_args, algorithm: BaseAlgorithm = None, **rule_kwargs) -> list[int]:
